changed structure of rl_glue folder to better separate areas of code
