+++ rl_glue/environments/environment.py
@@ -1,0 +1,74 @@
+#!/usr/bin/env python
+
+"""
+  Author: Adam White, Mohammad M. Ajallooeian
+  Purpose: for use of Reinforcement learning course University of Alberta Fall 2017
+
+  env *ignores* actions: rewards are all random
+"""
+
+from __future__ import print_function
+
+from abc import ABCMeta, abstractmethod
+
+
+class Environment:
+    """Implements the environment for an RLGlue environment
+
+    Note:
+        env_init, env_start, env_step, env_cleanup, and env_message are required
+        methods.
+    """
+
+    __metaclass__ = ABCMeta
+
+    def __init__(self):
+        reward = None
+        observation = None
+        termination = None
+        self.reward_obs_term = (reward, observation, termination)
+
+    @abstractmethod
+    def env_init(self, env_info={}):
+        """Setup for the environment called when the experiment first starts.
+
+        Note:
+            Initialize a tuple with the reward, first state observation, boolean
+            indicating if it's terminal.
+        """
+
+    @abstractmethod
+    def env_start(self):
+        """The first method called when the experiment starts, called before the
+        agent starts.
+
+        Returns:
+            The first state observation from the environment.
+        """
+
+    @abstractmethod
+    def env_step(self, action):
+        """A step taken by the environment.
+
+        Args:
+            action: The action taken by the agent
+
+        Returns:
+            (float, state, Boolean): a tuple of the reward, state observation,
+                and boolean indicating if it's terminal.
+        """
+
+    @abstractmethod
+    def env_cleanup(self):
+        """Cleanup done after the environment ends"""
+
+    @abstractmethod
+    def env_message(self, message):
+        """A message asking the environment for information
+
+        Args:
+            message: the message passed to the environment
+
+        Returns:
+            the response (or answer) to the message
+        """
